Add alias to approle secret IDs
When using a single approle role_id with multiple secret_ids, each
carrying different metadata, the latest authenticated secret id will
define the identity alias for that role_id. This means that all earlier
obtained tokens for the same role_id will use the metadata of the latest
created token in templating (e.g. policy templating), breaking the
scenario of a single role_id with multiple secret_ids to hand out to
different clients that need different permissions.

We fix this by adding an alias field to secret_id, which should contain
a unique value for each "end client", and which is used to define the
identy alias. I.e. in general secret IDs with different metadata should
have different aliases. Other options would be to use the secret
accessor id as identity alias, or to put simply a hash of the metadata,
but we opted for this solution to have a limit on the entries in the
identity alias table - we would flood it over time if secret IDs change
rapidly or a metadata field like a timestamp would be added.

Signed-off-by: Peter Verraedt <peter@verraedt.be>